move tests to public folder
