+++ sdk/communication/communication-phone-numbers/test/public/headers.spec.ts
@@ -7,11 +7,11 @@ import { DefaultAzureCredential } from "@azure/identity";
 import { isPlaybackMode } from "@azure/test-utils-recorder";
 import { assert } from "chai";
 import sinon from "sinon";
-import { PhoneNumbersClient } from "../src/phoneNumbersClient";
-import { getPhoneNumberHttpClient } from "./utils/mockHttpClients";
-import { SDK_VERSION } from "../src/utils/constants";
+import { PhoneNumbersClient } from "../../src";
+import { getPhoneNumberHttpClient } from "../utils/mockHttpClients";
+import { SDK_VERSION } from "../../src/utils/constants";
 import { Context } from "mocha";
-import { canCreateRecordedClientWithToken } from "./utils/recordedClient";
+import { canCreateRecordedClientWithToken } from "../utils/recordedClient";
 
 describe("PhoneNumbersClient - headers", function() {
   const endpoint = "https://contoso.spool.azure.local";
