+++ sdk/communication/communication-phone-numbers/tests.yml
@@ -5,8 +5,6 @@ stages:
     parameters:
       PackageName: "@azure/communication-phone-numbers"
       ServiceDirectory: communication
-      MatrixFilters:
-        - DependencyVersion=^$
       EnvVars:
         AZURE_SUBSCRIPTION_ID: $(acs-subscription-id)
         AZURE_COMMUNICATION_LIVETEST_CONNECTION_STRING: $(communication-livetest-connection-string)
